\name{cc.505}
\alias{cc.505}
\docType{data}
\title{
Select Biomarkers from the HVTN 505 Correlates Analysis
}
\description{
See reference.
}
\usage{data("cc.505")}
\format{
  A data frame with 189 observations on the following 27 variables.
  \describe{
    \item{\code{ptid}}{a character vector}
    \item{\code{trt}}{a numeric vector}
    \item{\code{case}}{a numeric vector}
    \item{\code{control}}{a numeric vector}
    \item{\code{perprot}}{a numeric vector}
    \item{\code{last_uninfec_immun_vst}}{a numeric vector}
    \item{\code{racefull}}{a numeric vector}
    \item{\code{racefulltxt}}{a character vector}
    \item{\code{bmi}}{a numeric vector}
    \item{\code{bmicat}}{a numeric vector}
    \item{\code{bmicattxt}}{a character vector}
    \item{\code{earliest_pos_vst}}{a numeric vector}
    \item{\code{level}}{a character vector}
    \item{\code{matchlevel}}{a character vector}
    \item{\code{samplingfraction}}{a numeric vector}
    \item{\code{vst9subcohort}}{a numeric vector}
    \item{\code{HIVwk28preunbl}}{a numeric vector}
    \item{\code{age}}{a numeric vector}
    \item{\code{racecc}}{a character vector}
    \item{\code{bhvrisk}}{a numeric vector}
    \item{\code{BMI}}{a numeric vector}
    \item{\code{stratuminds_vaccs}}{a numeric vector}
    \item{\code{stratuminds}}{a numeric vector}
    \item{\code{cd4.env.poly}}{a numeric vector}
    \item{\code{cd8.env.poly}}{a numeric vector}
    \item{\code{mfounders}}{a numeric vector}
    \item{\code{wei}}{a numeric vector}
  }
}
\references{
Fong, Y, Xu, J. Multi-Stage Simultaneous Deep Autoencoder-based Monotone (MSS-DAM) Nonlinear Dimensionality Reduction Methods, Journal of Computational and Graphical Statistics, in press.
}
