\name{hvtn505tier1}
\alias{hvtn505tier1}
\docType{data}
\title{
HVTN 505 Immune Correlates Tier 1 Dataset
}
\description{Contains eight immune response variables from the vaccine arm of the HVTN 505 trial.
}
\usage{data("hvtn505tier1")}
\format{
  A data frame with 150 observations on the following 9 variables.
  \describe{
    \item{\code{ptid}}{a character vector}
    \item{\code{CD8_ANYVRCENV_PolyfunctionalityScore_score}}{a numeric vector}
    \item{\code{IgGw28_env_mdw}}{a numeric vector}
    \item{\code{IgGw28_V1V2_mdw}}{a numeric vector}
    \item{\code{IgGw28_gp41_mdw}}{a numeric vector}
    \item{\code{ADCP1}}{a numeric vector}
    \item{\code{R2aConSgp140CFI}}{a numeric vector}
    \item{\code{IgAw28_env_mdw}}{a numeric vector}
    \item{\code{IgG3w28_env_mdw}}{a numeric vector}
  }
}
\references{
Fong, Y, Xu, J. Multi-Stage Simultaneous Deep Autoencoder-based Monotone (MSS-DAM) Nonlinear Dimensionality Reduction Methods, Journal of Computational and Graphical Statistics, in press.

Janes, H.E., Cohen, K.W., Frahm, N., De Rosa, S.C., Sanchez, B., Hural, J. et al (2017), Higher T-cell responses induced by DNA/rAd5 HIV-1 preventive vaccine are associated with lower HIV-1 infection risk in an efficacy trial, The Journal of infectious diseases, 215, 1376-1385.

Fong, Y., Shen, X., Ashley, V.C., Deal, A., Seaton, K.E., Yu, C. et al (2018), Vaccine-induced antibody responses modify the association between T-cell immune responses and HIV-1 infection risk in HVTN 505, The Journal of Infectious Diseases, 217, 1280--1288. 

Neidich, S.D., Fong, Y., Shen, X., Ashley, V.C., Deal, A., Seaton, K.E. et al (2019), Antibody Fc-effector Functions and IgG3 Associates with Decreased HIV-1 Acquisition Risk, The Journal of Infectious Diseases, 129, 4838-4849.
}
